refactor: chat room connection 폴더 구분 및 리팩터링 수행
