+++ src/main/java/com/sptp/backend/chat_room/service/ChatRoomService.java
@@ -3,8 +3,8 @@ package com.sptp.backend.chat_room.service;
 import com.sptp.backend.art_work.repository.ArtWorkRepository;
 import com.sptp.backend.aws.service.AwsService;
 import com.sptp.backend.chat_room.repository.ChatRoom;
-import com.sptp.backend.chat_room.repository.ChatRoomConnection;
-import com.sptp.backend.chat_room.repository.ChatRoomConnectionRepository;
+import com.sptp.backend.chat_room_connection.repository.ChatRoomConnection;
+import com.sptp.backend.chat_room_connection.repository.ChatRoomConnectionRepository;
 import com.sptp.backend.chat_room.repository.ChatRoomRepository;
 import com.sptp.backend.chat_room.web.dto.ChatRoomDetailResponse;
 import com.sptp.backend.chat_room.web.dto.ChatRoomResponse;
@@ -28,7 +28,6 @@ import java.util.stream.Collectors;
 public class ChatRoomService {
 
     private final ChatRoomRepository chatRoomRepository;
-    private final ChatRoomConnectionRepository chatRoomConnectionRepository;
     private final MemberRepository memberRepository;
     private final ArtWorkRepository artWorkRepository;
     private final MessageRepository messageRepository;
@@ -106,24 +105,6 @@ public class ChatRoomService {
 
         if (chatRoomRepository.existsById(chatRoomId)) {
             chatRoomRepository.deleteById(chatRoomId);
-        }
-    }
-
-    public void connect(String sessionId, Long chatRoomId, Long memberId) {
-
-        ChatRoomConnection chatRoomConnection = ChatRoomConnection.builder()
-                .sessionId(sessionId)
-                .chatRoomId(chatRoomId)
-                .memberId(memberId)
-                .build();
-
-        chatRoomConnectionRepository.save(chatRoomConnection);
-    }
-
-    public void disconnect(String sessionId) {
-
-        if (chatRoomConnectionRepository.existsBySessionId(sessionId)) {
-            chatRoomConnectionRepository.deleteBySessionId(sessionId);
         }
     }
 }
