+++ src/test/java/com/sptp/backend/message/service/MessageServiceTest.java
@@ -2,7 +2,7 @@ package com.sptp.backend.message.service;
 
 import com.sptp.backend.art_work.repository.ArtWork;
 import com.sptp.backend.chat_room.repository.ChatRoom;
-import com.sptp.backend.chat_room.repository.ChatRoomConnectionRepository;
+import com.sptp.backend.chat_room_connection.repository.ChatRoomConnectionRepository;
 import com.sptp.backend.chat_room.repository.ChatRoomRepository;
 import com.sptp.backend.common.exception.CustomException;
 import com.sptp.backend.common.exception.ErrorCode;
